requirements.txt
litellm
chromadb
python-dotenv
rich
